microservice/requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
flake8==6.0.0
gunicorn==21.2.0
mccabe==0.7.0
packaging==23.1
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
